Restructured DOM for better behavior
Changed layout: there is a left flex container and a right flex container which change size to center the article. Now the sidebar clings to the side of the article properly on larger page widths.

Also did some minor styling changes to the bar benear the title and the text in subtitles.